Add files via upload
This Jupyter Notebook file contains Python code to explore the relationship between sampling distribution vs the population, namely it generates a list of numbers (either of your specification or by random), calculates the sampling means of sample sizes of your choice, and compare the result to the population mean. It also generates a histogram for fun. Codes are almost exclusively written without using any library for fun.